Add form prefixes
This adds form prefixes to both forms, meaning that when they appear on
a page with other forms then validation doesn't get performed on the
wrong form.

For example, if a form error shows on a page for another form, the
mailing list form gets "This field is required" as an error, something
that's confusing if the user isn't trying to fill that form out.